Omnisearch code moved to search.ts
